IOU Object Image comparator
This script will rotate the object by 45 degrees rotations, checking each render of said rotation with the original image of the object and calculate it's respective IOU, printing each angle's IOU and the maximum IOU that was calculated with it's respective angle. 

The image and object paths will need to be changed per usage, as well as a folder / directory for saved renders for comparison.